tests/test_requirements.txt
###############################
## Dependencies for test functions
## call this file to install all dependencies
## 		> pip install -r requirements.txt
###############################

numpy>=1.13.1
pydicom>=1.0.2
nibabel>=2.1.0
pyzmq>=16.0.2
pyyaml>=3.12
